% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\name{cosangmx}
\alias{cosangmx}
\title{Get cosinus of pairwise angles between data points and interpolation point as fixed point}
\usage{
cosangmx(data, xinp, distance = NULL)
}
\arguments{
\item{data}{A matrix of data points}

\item{xinp}{Coordinates of the fixed point of the angles; usually the point where interpolation has to be computed}

\item{distance}{Distances between xinp and data can be given as an optional argument, if they have been coputed before, such that a new computation is not necessary}
}
\value{
The cosini of the pairwise angles between two data points with fixed point xinp
}
\description{
This function is basically a helper function for mvInterpolation::ti_get and not intended for direct use.
}
\examples{
coords <- matrix(runif(10*4)*10, ncol=4)
cosangmx(coords,rep(4,4))
}
