% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\name{di_get}
\alias{di_get}
\title{Get weights for interpolation by inverse distance weighting}
\usage{
di_get(data, xinp, p = 2)
}
\arguments{
\item{data}{A matrix of the points for which the function is known}

\item{xinp}{Coordinates, where interpolation has to be computed}

\item{p}{additional parameter as in the Shepard paper}
}
\value{
The weights for the IDW interpolation
}
\description{
Interpolation weights for inverse distance weighting.
}
\examples{
mydata <- matrix(runif(10*4)*10, ncol=4)
mydata <- cbind(mydata,abs(apply(mydata,1,sum)-3),abs(apply(mydata,1,prod)-4))
di_get(mydata, c(4,4,4,4))
}
