% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\name{getR}
\alias{getR}
\title{Get initial search-radius}
\usage{
getR(dat, d)
}
\arguments{
\item{dat}{A matrix of the points for which the function is known}

\item{d}{Dimension of the coordinate space}
}
\value{
The radius as a  numeric
}
\description{
This function determines a search radius such that on average 7 points lie
within the ball with the radius.
}
