% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\docType{package}
\name{mvInterpolation-package}
\alias{mvInterpolation}
\alias{mvInterpolation-package}
\title{Generate R documentation from inline comments.}
\description{
Roxygen2 allows you to write documentation in comment blocks co-located
with code.
}
\details{
The only function Otherwise refer to the vignettes to see
how to format the documentation.
}
\references{
Shepard, Donald (1968). "A two-dimensional interpolation function for irregularly-spaced data". Proceedings of the 1968 ACM National Conference. pp. 517–524. doi:10.1145/800186.810616.
}
\author{
Andreas Mändle, \email{maendle@uni-bremen.de}
}
\keyword{interpolation}
\keyword{multivariate}
