% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\name{slopezi_get}
\alias{slopezi_get}
\title{Get drift based on slopes}
\usage{
slopezi_get(data, xinp, rad = getR, vparam = 0.1)
}
\arguments{
\item{data}{A matrix of the points for which the function is known}

\item{xinp}{Coordinates, where interpolation has to be computed}

\item{rad}{Function which gives the initial radius for the observations considered in the interpolation}

\item{vparam}{Scalar, parameter for the slopes}
}
\value{
The weights for the ADW interpolation considering slopes and based on the nearest neighbors
}
\description{
The drift is needed for interpolation by angular distance weighting extended by slopes.
Interpolation weights are based on the nearest neighbors (4 to 10 neighbors, depending on the distance).
}
\examples{
mydata <- matrix(runif(10*4)*10, ncol=4)
mydata <- cbind(mydata,abs(apply(mydata,1,sum)-3),abs(apply(mydata,1,prod)-4))
slopezi_get(mydata, c(4,4,4,4))
}
