% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shepard.R
\name{ti_get}
\alias{ti_get}
\title{Get weights for interpolation by angular distance weighting (ADW)}
\usage{
ti_get(data, xinp, rad = getR)
}
\arguments{
\item{data}{A matrix of the points for which the function is known}

\item{xinp}{Coordinates, where interpolation has to be computed}

\item{rad}{Function which gives the initial radius for the observations considered in the interpolation}
}
\value{
The weights for the ADW interpolation based on the nearest neighbors
}
\description{
Interpolation weights for angular distance weighting of the nearest neighbors (4 to 10 neighbors, depending on the distance)
}
\examples{
mydata <- matrix(runif(10*4)*10, ncol=4)
mydata <- cbind(mydata,abs(apply(mydata,1,sum)-3),abs(apply(mydata,1,prod)-4))
ti_get(mydata, c(4,4,4,4))
}
